Create bugfix branch for 0.4.1.

+++ go/GoNodeUtil.cpp
@@ -34,8 +34,7 @@ GoKomi GoNodeUtil::GetKomi(const SgNode* node)
         {
             try
             {
-                return GoKomi(static_cast<float>
-                              (node->GetRealProp(SG_PROP_KOMI)));
+                return GoKomi(node->GetRealProp(SG_PROP_KOMI));
             }
             catch (const GoKomi::InvalidKomi&)
             {
